Add uniform resampling
This is a major change on the pipeline: at the very beginning of the galvanic
clean signal, resample the signal to have a uniform sample rate. Many of our
algorithms that occur in this step or downstream from this step assume a uniform
step size.

Also, added more verbosity on the CleanSignal task